+++ tests/testthat/test-array.R
@@ -6,7 +6,8 @@ withr::local_file(parentDirArr)
 test_that('checkSeekerArrayArgs', {
   skip_on_os('windows', arch = NULL)
 
-  resultObs = checkSeekerArrayArgs(paramsArray$study, paramsArray$geneIdType, paramsArray$platform, parentDirArr)
+  resultObs = checkSeekerArrayArgs(paramsArray$study, paramsArray$geneIdType,
+                                   paramsArray$platform, parentDirArr)
   resultExp = snapshot(
     resultObs, file.path(dataDir, 'seeker_array_args_output.qs'))
 
@@ -17,19 +18,23 @@ test_that('checkSeekerArrayArgs errors', {
   skip_on_os('windows', arch = NULL)
 
   # GSE platform not null or GPL
-  expect_error(checkSeekerArrayArgs(paramsArray$study, paramsArray$geneIdType, 'abcd', parentDirArr))
+  expect_error(checkSeekerArrayArgs(paramsArray$study, paramsArray$geneIdType,
+                                    'abcd', parentDirArr))
 
   # E- with platform
-  expect_error(checkSeekerArrayArgs('E-test', paramsArray$geneIdType, 'abcd', parentDirArr))
+  expect_error(checkSeekerArrayArgs('E-test', paramsArray$geneIdType, 'abcd',
+                                    parentDirArr))
 
   # raw with raw dir not existing
-  expect_error(checkSeekerArrayArgs('LOCAL', paramsArray$geneIdType, 'GPL1', parentDirArr))
+  expect_error(checkSeekerArrayArgs('LOCAL', paramsArray$geneIdType, 'GPL1',
+                                    parentDirArr))
 })
 
 test_that('seekerArray GSE', {
   skip_on_os('windows', arch = NULL)
 
-  seekerArray(params = paramsArray, parentDir = parentDirArr)
+  seekerArray(paramsArray$study, paramsArray$geneIdType, paramsArray$platform,
+              parentDir = parentDirArr)
 
   resultObs = list.files(parentDirArr, recursive = TRUE)
   resultExp = snapshot(
@@ -47,7 +52,8 @@ test_that('seekerArray Ae', {
   paramsArrayAe = paramsArray
   paramsArrayAe$study = 'E-MTAB-8714'
 
-  seekerArray('E-MTAB-8714', paramsArray$geneIdType, paramsArray$platform, parentDir = parentDirArrAe)
+  seekerArray(paramsArrayAe$study, paramsArrayAe$geneIdType,
+              paramsArrayAe$platform, parentDir = parentDirArrAe)
 
   resultObs = list.files(parentDirArrAe, recursive = TRUE)
   resultExp = snapshot(
@@ -65,7 +71,8 @@ test_that('seekerArray LOCAL', {
   paramsArrayLocal = yaml::read_yaml(file.path(dataDir, 'LOCAL01.yml'))
   file.copy(file.path(dataDir, 'LOCAL01'), parentDirArrLcl, recursive = TRUE)
 
-  seekerArray(params = paramsArrayLocal, parentDir = parentDirArrLcl)
+  seekerArray(paramsArrayLocal$study, paramsArrayLocal$geneIdType,
+              paramsArrayLocal$platform, parentDir = parentDirArrLcl)
 
   resultObs = list.files(parentDirArrLcl, recursive = TRUE)
   resultExp = snapshot(