+++ NEWS.md
@@ -1,6 +1,6 @@
 # seeker 1.0.3
-* Make `getProbeGeneMapping` test check for version number of mapping package.
-* Revised arguments for `seekerArray` and updated tests/vignettes accordingly.
+* Made `getProbeGeneMapping()` test check for version number of mapping package.
+* Revised arguments for `seekerArray()` and updated tests/vignettes accordingly.
 
 # seeker 1.0.2
 * Set default parent directory for `seekerArray()`.